added new pages

+++ src/views/Admins/Admins/index.tsx
@@ -1,4 +1,4 @@
-import { useCallback } from "react";
+import { useCallback, useState } from "react";
 import CTable from "../../../components/CElements/CTable";
 import AddButton from "../../../components/UI/Buttons/AddButton";
 import SectionHeader from "../../../components/UI/Sections/Header";
@@ -8,6 +8,7 @@ import adminService from "../../../services/admins";
 import { Header } from "../../../components/UI/Header";
 import CBreadcrumbs from "../../../components/CElements/CBreadcrumbs";
 import { FetchFunction, TableData, breadCrumbs } from "./Logic";
+import { FilterFunctions } from "../../../components/UI/Filter/Logic";
 
 const Admins = () => {
   const { navigateQuery } = usePageRouter();
@@ -15,8 +16,17 @@ const Admins = () => {
   const query = getQueries();
   const { headColumns } = TableData();
   const { bodyColumns, isLoading, refetch } = FetchFunction();
+  const [filterParams, setFilterParams]: any = useState({});
+  const { collectFilter, storeFilters } = FilterFunctions({ store: true, filterParams, setFilterParams });
 
-  const handleSearch = () => {};
+  const handleSearch = (value: any) => {
+    collectFilter({ type: "q", val: value });
+  };
+
+  const handleFilterParams = (obj: any) => {
+    setFilterParams(obj);
+    storeFilters(obj);
+  };
 
   const handleActions = useCallback((element: any, status: string) => {
     if (status === "view") navigateQuery({ id: element.id });
@@ -24,6 +34,7 @@ const Admins = () => {
     if (status === "edit") navigateQuery({ id: element.id });
 
     if (status === "delete") {
+      if (element.email === 'superrt@akataxi.uz') return
       adminService.deleteAdmin(element.id).then(() => {
         refetch();
       });
@@ -36,7 +47,7 @@ const Admins = () => {
         <CBreadcrumbs items={breadCrumbs} progmatic={true} />
       </Header>
       <div className="container">
-        <SectionHeader handleSearch={handleSearch}>
+        <SectionHeader handleSearch={handleSearch} defaultValue={filterParams?.q}>
           <div className="flex items-center gap-3">
             <AddButton
               text="new_admin"
@@ -52,6 +63,8 @@ const Admins = () => {
           isResizeble={true}
           isLoading={isLoading}
           handleActions={handleActions}
+          filterParams={filterParams}
+          handleFilterParams={handleFilterParams}
         />
 
         {query?.id && <Form refetch={refetch} id={query.id} />}
